requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
crispy-bootstrap5==0.7
dj-database-url==1.3.0
Django==4.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.10.0
django-filter==22.1
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
Pillow==9.2.0
psycopg==3.1.8
psycopg2-binary==2.9.6
pycodestyle==2.9.1
python-magic==0.4.27
pytz==2022.6
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.15
yapf==0.32.0
